update & refactoring: test기능 세분화 및 tokenizer test코드 추가
